pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medical-diagnostic-env"
version = "1.0.0"
description = "OpenEnv environment for medical diagnosis RL training"
readme = "README.md"
requires-python = ">=3.10"

authors = [
    {name = "Team SYNAPSE", email = "synapse@example.com"}
]

keywords = [
    "reinforcement-learning",
    "medical",
    "diagnosis",
    "healthcare",
    "rl-training",
    "llm",
    "openenv"
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openenv-core>=0.2.3",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=16.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "openai>=1.3.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/meta-pytorch/OpenEnv"
Documentation = "https://meta-pytorch.org/OpenEnv/"
Repository = "https://github.com/meta-pytorch/OpenEnv"
Issues = "https://github.com/meta-pytorch/OpenEnv/issues"

[tool.setuptools]
packages = ["medical_diagnostic_env"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_untyped_defs = false
warn_unused_ignores = true
